Feat: UserService

+++ src/controllers/UserController.ts
@@ -7,7 +7,7 @@ export class UserController {
 
     async getAllUsers(request: Request, response: Response) {
 
-        const users = userService.getAllUsers()
+        const users = await userService.getAllUsers()
 
         return response.status(200).json(users)
 
@@ -16,6 +16,10 @@ export class UserController {
     async createUser(request: Request, response: Response) {
 
         const user = request.body
+
+        if(!user.username || !user.password) {
+            return response.status(422).json({ message: "Invalid informations!" })
+        }
 
         await userService.createUser(user.username, user.password)
 
@@ -28,6 +32,10 @@ export class UserController {
         const user = request.body
         const userId = request.params.id
 
+        if(!user.username || !user.password || !userId) {
+            return response.status(422).json({ message: "Invalid informations!" })
+        }
+
         await userService.updateUser(userId, user.username, user.password)
 
         return response.status(200).json({ message: "User updated!" })
@@ -37,6 +45,10 @@ export class UserController {
     async deleteUser(request: Request, response: Response) {
 
         const userId = request.params.id
+
+        if(!userId) {
+            return response.status(422).json({ message: "Invalid information!" })
+        }
 
         await userService.deleteUser(userId)
